requirements.txt
python-dotenv
Pillow
pyrogram
tgcrypto
